Experiment 5:
I realize now that the issue is NOT high variance like I thought
in the last experiment. With the addition of training metrics, I
can see that the model just has high bias and isnt learning well
at all.

I'm going to try to remove some of the regularization and just
get back to the model overfitting the training set like we had
in experiment 1. Then regularize from there.