Experiment 11
Same params/model as experiment 10, except I trained it from
scratch instead of on a pretrained experiment. Performance was
best seen so far, with an MS-SSIM of 0.46.

However, around 200 epochs in, the discriminator got way better
than the generator and the generator rapidly diverged and got
stuck outputting random noise

I plan to rerun this trial but with Wasserstein loss to see how
much better it can be.