Experiment 12
First experiment with WassersteinGAN with Gradient Penalty loss.
Some of the best performance so far - but in a gimmicky way. The
generator ended up learning to output the same image for every
input.

However this loss function was pretty good for making sure that
the gradients didn't completely vanish like last experiment. I
think some tuning of this setup could lead to some good results.